+++ src/components/MainContainer.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import MainTopbar from './MainTopbar'
 import ApplicationContainer from './ApplicationContainer'
-import DomainsContainer from './DomainsContainer'
-import DomainProfile from './DomainProfile'
+import DomainsContainer from './domains/DomainsContainer'
+import DomainProfile from './single_domain/DomainProfile'
 import AccountDashboard from './dashboard/AccountDashboard'
 import RocketChat from './RocketChat'
 import GovernanceContainer from './governance/GovernanceContainer'
