+++ src/components/domains/DomainsContainer.js
@@ -4,20 +4,20 @@ import qs from 'qs'
 import updateQuery from 'update-query'
 // import { Popup } from 'semantic-ui-react'
 
-import normalizeQueryObj from '../utils/normalizeQueryObj'
+import normalizeQueryObj from '../../utils/normalizeQueryObj'
 
-import RegistryStatsbar from './RegistryStatsbar'
-import AdtCalculator from './AdtCalculator'
+import RegistryStatsbar from '../RegistryStatsbar'
+import AdtCalculator from '../AdtCalculator'
 import DomainsTable from './DomainsTable'
 import DomainsFilterPanel from './DomainsFilterPanel'
 // import Trollbox from './Trollbox'
-import WelcomeModal from './WelcomeModal'
-import RegistryGuideStaticChallenge from './registry_guide/RegistryGuideStaticChallenge'
-import RegistryGuideStaticInRegistry from './registry_guide/RegistryGuideStaticInRegistry'
-import RegistryGuideStaticVoting from './registry_guide/RegistryGuideStaticVoting'
-import RegistryGuideStaticReveal from './registry_guide/RegistryGuideStaticReveal'
-import RegistryGuideStaticDashboard from './registry_guide/RegistryGuideStaticDashboard'
-import DomainEmailNotifications from './DomainEmailNotifications'
+import WelcomeModal from '../WelcomeModal'
+import RegistryGuideStaticChallenge from '../registry_guide/RegistryGuideStaticChallenge'
+import RegistryGuideStaticInRegistry from '../registry_guide/RegistryGuideStaticInRegistry'
+import RegistryGuideStaticVoting from '../registry_guide/RegistryGuideStaticVoting'
+import RegistryGuideStaticReveal from '../registry_guide/RegistryGuideStaticReveal'
+import RegistryGuideStaticDashboard from '../registry_guide/RegistryGuideStaticDashboard'
+import DomainEmailNotifications from '../DomainEmailNotifications'
 
 import './DomainsContainer.css'
