+++ src/components/single_domain/DomainChallengeContainer.js
@@ -4,12 +4,12 @@ import commafy from 'commafy'
 import toastr from 'toastr'
 import moment from 'moment'
 import { Button } from 'semantic-ui-react'
-import Tooltip from './Tooltip'
-import calculateGas from '../utils/calculateGas'
+import Tooltip from '../Tooltip'
+import calculateGas from '../../utils/calculateGas'
 
-import Countdown from './CountdownText'
-import registry from '../services/registry'
-import parametizer from '../services/parameterizer'
+import Countdown from '../CountdownText'
+import registry from '../../services/registry'
+import parametizer from '../../services/parameterizer'
 import DomainChallengeInProgressContainer from './DomainChallengeInProgressContainer'
 import PubSub from 'pubsub-js'
 import Eth from 'ethjs'
