+++ src/components/single_domain/DomainChallengeInProgressContainer.css
