+++ src/components/single_domain/DomainChallengeInProgressContainer.js
