+++ src/components/single_domain/DomainInRegistryContainer.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types'
 import toastr from 'toastr'
 import { Button, Input, Segment } from 'semantic-ui-react'
 import commafy from 'commafy'
-import Tooltip from './Tooltip'
-import calculateGas from '../utils/calculateGas'
+import Tooltip from '../Tooltip'
+import calculateGas from '../../utils/calculateGas'
 
-import registry from '../services/registry'
+import registry from '../../services/registry'
 import './DomainInRegistryContainer.css'
 import DomainChallengeInProgressContainer from './DomainChallengeInProgressContainer'
-import WithdrawInProgressContainer from './dashboard/WithdrawInProgressContainer'
+import WithdrawInProgressContainer from '../dashboard/WithdrawInProgressContainer'
 import TopOffInProgressContainer from './TopOffInProgressContainer'
 import DomainChallengeContainer from './DomainChallengeContainer'
 import Eth from 'ethjs'
