+++ src/components/single_domain/DomainProfile.js
@@ -5,7 +5,7 @@ import qs from 'qs'
 import DomainProfileHeader from './DomainProfileHeader'
 import DomainStatsbar from './DomainStatsbar'
 // import DomainScamReport from './DomainScamReport'
-import DomainRedditBox from './reddit/DomainRedditBox'
+import DomainRedditBox from '../reddit/DomainRedditBox'
 import DomainProfileActionContainer from './DomainProfileActionContainer'
 // import DomainProfileAdsTxtStatus from './DomainProfileAdsTxtStatus'
 import DomainProfileStageMap from './DomainProfileStageMap'
