+++ src/components/single_domain/DomainProfileHeader.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Tooltip from './Tooltip'
+import Tooltip from '../Tooltip'
 import './DomainProfileHeader.css'
 
 function DomainProfileHeader (props) {
