+++ src/components/single_domain/DomainProfileStageMap.js
@@ -2,20 +2,20 @@ import React, { Component } from 'react'
 import toastr from 'toastr'
 
 import './DomainProfileStageMap.css'
-import Tooltip from './Tooltip'
-import getDomainState from '../utils/determineDomainState'
+import Tooltip from '../Tooltip'
+import getDomainState from '../../utils/determineDomainState'
 
-import MapInApplication from './assets/stage_maps/map_application.svg'
-import MapInApplicationPending from './assets/stage_maps/map_in_application_pending.svg'
-import MapInRegistryNoChallengeId from './assets/stage_maps/map_in_registry_no_id.svg'
-import MapCommit from './assets/stage_maps/map_commit.svg'
-import MapReveal from './assets/stage_maps/map_reveal.svg'
-import MapRevealPending from './assets/stage_maps/map_reveal_pending.svg'
-import MapRejected from './assets/stage_maps/map_rejected.svg'
-import MapInRegistryChallengeId from './assets/stage_maps/map_in_registry_challenge_id.svg'
-import MapInRegistryCommit from './assets/stage_maps/map_in_registry_challenge_id_commit.svg'
-import MapInRegistryReveal from './assets/stage_maps/map_in_registry_no_id_reveal.svg'
-import MapInRegistryRevealPending from './assets/stage_maps/map_in_registry_no_id_reveal_pending.svg'
+import MapInApplication from '../assets/stage_maps/map_application.svg'
+import MapInApplicationPending from '../assets/stage_maps/map_in_application_pending.svg'
+import MapInRegistryNoChallengeId from '../assets/stage_maps/map_in_registry_no_id.svg'
+import MapCommit from '../assets/stage_maps/map_commit.svg'
+import MapReveal from '../assets/stage_maps/map_reveal.svg'
+import MapRevealPending from '../assets/stage_maps/map_reveal_pending.svg'
+import MapRejected from '../assets/stage_maps/map_rejected.svg'
+import MapInRegistryChallengeId from '../assets/stage_maps/map_in_registry_challenge_id.svg'
+import MapInRegistryCommit from '../assets/stage_maps/map_in_registry_challenge_id_commit.svg'
+import MapInRegistryReveal from '../assets/stage_maps/map_in_registry_no_id_reveal.svg'
+import MapInRegistryRevealPending from '../assets/stage_maps/map_in_registry_no_id_reveal_pending.svg'
 
 import PubSub from 'pubsub-js'
