+++ src/components/single_domain/DomainRejectedContainer.css
