+++ src/components/single_domain/DomainRejectedContainer.js
@@ -1,36 +1,32 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Button } from 'semantic-ui-react'
-import Tooltip from './Tooltip'
-import './DomainPendingContainer.css'
-import registry from '../services/registry'
+import Tooltip from '../Tooltip'
+import './DomainRejectedContainer.css'
+import registry from '../../services/registry'
 import toastr from 'toastr'
-import RefreshInProgressContainer from './RefreshInProgressContainer'
-import PubSub from 'pubsub-js'
 
-class DomainPendingContainer extends Component {
+class DomainRejectedContainer extends Component {
   constructor (props) {
     super()
 
     this.state = {
-      domain: props.domain,
-      inProgress: false
+      domain: props.domain
     }
   }
 
   render () {
     const {
-      domain,
-      inProgress
+      domain
     } = this.state
 
     return (
-      <div className='DomainPendingContainer'>
-        <div className='ui grid stackable DomainPendingBody'>
+      <div className='DomainRejectedContainer'>
+        <div className='ui grid stackable DomainRejectedBody'>
           <div className='column sixteen wide HeaderColumn'>
             <div className='row HeaderRow'>
               <div className='ui large header'>
-                Stage: Pending
+                Stage: Rejected
                 <Tooltip
                   info='The first phase of the voting process is the commit phase where the ADT holder stakes a hidden amount of votes to SUPPORT or OPPOSE the domain application. The second phase is the reveal phase where the ADT holder reveals the staked amount of votes to either the SUPPORT or OPPOSE side.'
                 />
@@ -46,38 +42,33 @@ class DomainPendingContainer extends Component {
             <div className='ui divider' />
           </div>
           <div className='column sixteen wide center aligned'>
-            <p className='PendingMessage'>
-              Please click on the <strong>REFRESH STATUS</strong> button above to refresh the correct stage for {domain}
+            <p className='RejectedMessage'>
+              You can apply <strong>{domain}</strong> into the
+              adChain Registry within the application modal
+              on the left navigation column.
             </p>
           </div>
         </div>
-        {inProgress ? <RefreshInProgressContainer /> : null}
       </div>
     )
   }
 
   async updateStatus (domain) {
-    this.setState({
-      inProgress: true
-    })
     try {
       await registry.updateStatus(domain)
-      await PubSub.publish('DomainProfileStageMap.updateStageMap')
-      this.setState({
-        inProgress: false
-      })
     } catch (error) {
-      console.error(error)
-      this.setState({
-        inProgress: false
-      })
-      toastr.error('There was an error updating domain')
+      try {
+        console.log(error)
+        toastr.error('There was an error updating domain')
+      } catch (err) {
+        console.log(err)
+      }
     }
   }
 }
 
-DomainPendingContainer.propTypes = {
+DomainRejectedContainer.propTypes = {
   domain: PropTypes.string
 }
 
-export default DomainPendingContainer
+export default DomainRejectedContainer
