+++ src/components/single_domain/DomainScamReport.js
