+++ src/components/single_domain/DomainVoteCommitContainer.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types'
 import toastr from 'toastr'
 import moment from 'moment'
 import { Input, Button, Segment } from 'semantic-ui-react'
-import Tooltip from './Tooltip'
+import Tooltip from '../Tooltip'
 import randomInt from 'random-int'
-import calculateGas from '../utils/calculateGas'
+import calculateGas from '../../utils/calculateGas'
 
-import saveFile from '../utils/saveFile'
-import Countdown from './CountdownText'
-import registry from '../services/registry'
+import saveFile from '../../utils/saveFile'
+import Countdown from '../CountdownText'
+import registry from '../../services/registry'
 import DomainVoteCommitInProgressContainer from './DomainVoteCommitInProgressContainer'
 
 import './DomainVoteCommitContainer.css'
