+++ src/components/single_domain/DomainVoteCommitInProgressContainer.js
