+++ src/components/single_domain/DomainVoteRevealContainer.css
