+++ src/components/single_domain/DomainVoteRevealInProgressContainer.css
