+++ src/components/single_domain/DomainVoteRevealInProgressContainer.js
